implement acl encode/decode
